requirements.txt
python-dotenv==1.0.1
pywhatkit==5.4
requests==2.32.3
schedule==1.2.2
